Fix tsd tests after `index.js` was moved into `src`
