+++ test-d/Components/ListComponent.test-d.ts
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import lng from '../../index';
+import lng from '../../src/index.js';
 import { expectType } from 'tsd';
 import TransitionSettings from '../../src/animation/TransitionSettings.mjs';
 import AnimationSettings from '../../src/animation/AnimationSettings.mjs';
