+++ test-d/Elements/Element.test-d.ts
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 import { expectType } from 'tsd';
-import lng from '../../index.js';
+import lng from '../../src/index.js';
 import ListComponent from '../../src/components/ListComponent.mjs';
 import { InlineElement, SmoothTemplate } from '../../src/tree/Element.mjs';
